--- a/CI/codespell_ignore.txt
+++ b/CI/codespell_ignore.txt
@@ -19,3 +19,4 @@ strack
 te
 vart
 writet
+matA