--- a/CI/codespell_ignore.txt
+++ b/CI/codespell_ignore.txt
@@ -19,4 +19,4 @@ strack
 te
 vart
 writet
-matA
+mata