--- a/CMakePresets.json
+++ b/CMakePresets.json
@@ -66,6 +66,7 @@
                 "ACTS_BUILD_PLUGIN_EXATRKX": "OFF",
                 "ACTS_BUILD_PLUGIN_JSON": "ON",
                 "ACTS_BUILD_PLUGIN_ONNX": "OFF",
+                "ACTS_BUILD_PLUGIN_ROOT": "ON",
                 "ACTS_BUILD_PLUGIN_TGEO": "ON",
                 "ACTS_BUILD_EXAMPLES_EDM4HEP": "ON",
                 "ACTS_BUILD_EXAMPLES_EXATRKX": "OFF",