--- a/Examples/Io/Root/CMakeLists.txt
+++ b/Examples/Io/Root/CMakeLists.txt
@@ -39,6 +39,7 @@ target_link_libraries(
     ActsExamplesIoRoot
     PUBLIC
         ActsCore
+        ActsPluginRoot
         ActsExamplesDigitization
         ActsExamplesFramework
         ActsExamplesPropagation