--- a/Examples/Io/Root/include/ActsExamples/Io/Root/RootMaterialTrackWriter.hpp
+++ b/Examples/Io/Root/include/ActsExamples/Io/Root/RootMaterialTrackWriter.hpp
@@ -8,11 +8,12 @@
 
 #pragma once
 
-#include "Acts/Definitions/Algebra.hpp"
-#include "ActsExamples/Framework/ProcessCode.hpp"
-#include "ActsExamples/Framework/WriterT.hpp"
+#include <Acts/Definitions/Algebra.hpp>
+#include <Acts/Plugins/Root/RootMaterialTrack.hpp>
 #include <Acts/Propagator/MaterialInteractor.hpp>
 #include <Acts/Utilities/Logger.hpp>
+#include <ActsExamples/Framework/ProcessCode.hpp>
+#include <ActsExamples/Framework/WriterT.hpp>
 
 #include <cstddef>
 #include <cstdint>
@@ -24,6 +25,11 @@
 
 class TFile;
 class TTree;
+
+namespace ROOT::Experimental {
+class RNTupleWriter;
+}  // namespace ROOT::Experimental
+
 namespace ActsExamples {
 struct AlgorithmContext;
 }  // namespace ActsExamples
@@ -51,7 +57,7 @@ class RootMaterialTrackWriter
     : public WriterT<
           std::unordered_map<std::size_t, Acts::RecordedMaterialTrack>> {
  public:
-  struct Config {
+  struct Config : public Acts::RootMaterialTrack::Config {
     /// material collection to write
     std::string inputMaterialTracks = "material-tracks";
     /// path of the output file
@@ -60,18 +66,8 @@ class RootMaterialTrackWriter
     std::string fileMode = "RECREATE";
     /// name of the output tree
     std::string treeName = "material-tracks";
-
-    /// Re-calculate total values from individual steps (for cross-checks)
-    bool recalculateTotals = false;
-    /// Write aut pre and post step (for G4), otherwise central step position
-    bool prePostStep = false;
-    /// Write the surface to which the material step correpond
-    bool storeSurface = false;
-    /// Write the volume to which the material step correpond
-    bool storeVolume = false;
-    /// Collapse consecutive interactions of a single surface into a single
-    /// interaction
-    bool collapseInteractions = false;
+    /// Write as RNTuple
+    bool rnTuple = false;
   };
 
   /// Constructor with
@@ -109,93 +105,10 @@ class RootMaterialTrackWriter
   TFile* m_outputFile = nullptr;
   /// The output tree name
   TTree* m_outputTree = nullptr;
-
-  /// Event identifier.
-  std::uint32_t m_eventId = 0;
-
-  /// start global x
-  float m_v_x = 0;
-  /// start global y
-  float m_v_y = 0;
-  /// start global z
-  float m_v_z = 0;
-  /// start global momentum x
-  float m_v_px = 0;
-  /// start global momentum y
-  float m_v_py = 0;
-  /// start global momentum z
-  float m_v_pz = 0;
-  /// start phi direction
-  float m_v_phi = 0;
-  /// start eta direction
-  float m_v_eta = 0;
-  /// thickness in X0/L0
-  float m_tX0 = 0;
-  /// thickness in X0/L0
-  float m_tL0 = 0;
-
-  /// step x (start) position (optional)
-  std::vector<float> m_step_sx;
-  /// step y (start) position (optional)
-  std::vector<float> m_step_sy;
-  /// step z (start) position (optional)
-  std::vector<float> m_step_sz;
-  /// step x position
-  std::vector<float> m_step_x;
-  /// step y position
-  std::vector<float> m_step_y;
-  /// step z position
-  std::vector<float> m_step_z;
-  /// step r position
-  std::vector<float> m_step_r;
-  /// step x (end) position (optional)
-  std::vector<float> m_step_ex;
-  /// step y (end) position (optional)
-  std::vector<float> m_step_ey;
-  /// step z (end) position (optional)
-  std::vector<float> m_step_ez;
-  /// step x direction
-  std::vector<float> m_step_dx;
-  /// step y direction
-  std::vector<float> m_step_dy;
-  /// step z direction
-  std::vector<float> m_step_dz;
-  /// step length
-  std::vector<float> m_step_length;
-  /// step material x0
-  std::vector<float> m_step_X0;
-  /// step material l0
-  std::vector<float> m_step_L0;
-  /// step material A
-  std::vector<float> m_step_A;
-  /// step material Z
-  std::vector<float> m_step_Z;
-  /// step material rho
-  std::vector<float> m_step_rho;
-
-  /// ID of the surface associated with the step
-  std::vector<std::uint64_t> m_sur_id;
-  /// Type of the surface associated with the step
-  std::vector<std::int32_t> m_sur_type;
-  /// x position of the surface intersection associated with the step
-  std::vector<float> m_sur_x;
-  /// y position of the surface intersection associated with the step
-  std::vector<float> m_sur_y;
-  /// z position of the surface intersection associated with the step
-  std::vector<float> m_sur_z;
-  /// r of the position of the surface intersection associated with the step
-  std::vector<float> m_sur_r;
-  /// the distance to the surface associated with the step
-  std::vector<float> m_sur_distance;
-  /// path correction when associating material to the given surface
-  std::vector<float> m_sur_pathCorrection;
-  /// Min range of the surface associated with the step
-  std::vector<float> m_sur_range_min;
-  /// Max range of the surface associated with the step
-  std::vector<float> m_sur_range_max;
-
-  /// ID of the volume associated with the step
-  std::vector<std::uint64_t> m_vol_id;
+  /// Experimental RNTuple writer
+  std::unique_ptr<ROOT::Experimental::RNTupleWriter> m_rntWriter = nullptr;
+  /// The root material track object
+  Acts::RootMaterialTrack m_rootMaterialTrack;
 };
 
 }  // namespace ActsExamples