--- a/Plugins/Root/include/Acts/Plugins/Root/RootMaterialTrack.hpp
+++ b/Plugins/Root/include/Acts/Plugins/Root/RootMaterialTrack.hpp
@@ -6,6 +6,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+#pragma once
+
 #include "Acts/Geometry/GeometryContext.hpp"
 #include "Acts/Material/MaterialInteraction.hpp"