This PR makes the Root based recording of Geant4 material (as input to the material mapping) to `Core` and puts reading and writing onto the same internal `payload`.

Summary: The new `RNTuple` backend results in a file size reduction of **35%**  and a per event runtime reduction (G4 recording and writing) of **17.5%**.

**Update: Unforunately I had to remove the RNTuple writing infrastructure for the moment as it is not yet stable thoughout our different ROOT versions, the code can be updated to it very quickly though.**


Details of recording a million material tracks in the ODD are as follows:

**(CURRENT MAIN) TTree backend:**

```sh
- Timing:
14:31:46    Sequencer      INFO      Processed 1000 events in 80.167658 s (wall clock)
14:31:46    Sequencer      INFO      Average time per event: 79.452257 ms/event

- File Size:
1154971403 21 Jan 14:31 geant4_material_tracks.root 
```

**(THIS PR) RNTuple backend:**

```sh
- Timing:
14:16:38    Sequencer      INFO      Processed 1000 events in 66.297680 s (wall clock)
14:16:38    Sequencer      INFO      Average time per event: 66.218186 ms/event

- File Size:
754161786 21 Jan 14:16 geant4_material_tracks_rntuple.root
```

**(THIS PR) TTree backend:**

```sh
- Timing:
14:28:30    Sequencer      INFO      Processed 1000 events in 81.241190 s (wall clock)
14:28:30    Sequencer      INFO      Average time per event: 80.542661 ms/event

- File Size:
1154970434 21 Jan 11:22 geant4_material_tracks_ttree.root
```


--- END COMMIT MESSAGE ---

Any further description goes here, @-mentions are ok here!

- Use a *conventional commits* prefix: [quick summary](https://www.conventionalcommits.org/en/v1.0.0/#summary)
  - We mostly use `feat`, `fix`, `refactor`, `docs`, `chore` and `build` types.
- A milestone will be assigned by one of the maintainers


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Release Notes

- **New Features**
  - Added support for ROOT RNTuple and expanded ROOT framework integration.
  - Enhanced material track reading and writing capabilities.
  - Introduced a new option to build the ROOT plugin.

- **Improvements**
  - Streamlined material track data handling.
  - Simplified configuration for ROOT-based material tracking.
  - Added new ROOT plugin component.

- **Technical Updates**
  - Updated CMake configurations to support new ROOT components.
  - Introduced `RootMaterialTrack` class for more efficient material track management.
  - Updated documentation to include new CMake option for ROOT plugin.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->